fix request body handling in lambda-proxy
Prior to this, the API Gateway simulator used various content type middlewares
for Express to parse request bodies, and then called JSON.stringify on the
parsed bodies before invoking the lambda. AWS just passes the body through
unchanged as a string in the event JSON. This commit matches AWS' behavior.